More warnings.

+++ code/rd-common/tr_image_jpg.cpp
@@ -513,7 +513,7 @@ size_t RE_SaveJPGToBuffer(byte *buffer, size_t bufSize, int quality,
 	/* Step 2: specify data destination (eg, a file) */
 	/* Note: steps 2 and 3 can be done in either order. */
 
-	jpegDest(&cinfo, buffer, bufSize);
+	jpegDest(&cinfo, buffer, static_cast<int>(bufSize));
 
 	/* Step 3: set parameters for compression */
 	cinfo.image_width = image_width; 	/* image width and height, in pixels */
@@ -575,10 +575,10 @@ void RE_SaveJPG(const char * filename, int quality, int image_width, int image_h
 	size_t bufSize;
 
 	bufSize = image_width * image_height * 3;
-	out = (byte *) R_Malloc( bufSize, TAG_TEMP_WORKSPACE, qfalse );
+	out = (byte *) R_Malloc( static_cast<int>(bufSize), TAG_TEMP_WORKSPACE, qfalse );
 
 	bufSize = RE_SaveJPGToBuffer(out, bufSize, quality, image_width, image_height, image_buffer, padding, false);
-	ri.FS_WriteFile(filename, out, bufSize);
+	ri.FS_WriteFile(filename, out, static_cast<int>(bufSize));
 
 	R_Free(out);
 }
